Интенсив AI/requirements.txt
requests
python-dotenv



